reorganize per folder sub components
