+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/ComplianceReadySoftwareSettings/complianceReadySettingsHelper.ts
