+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/PasswordInputField/__tests__/PasswordInputField.test.tsx
@@ -7,7 +7,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 
-import { PasswordInputField } from '../PasswordInputField'
+import { PasswordInputField } from '..'
 
 import type { ComponentProps } from 'react'
