+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/PersonalAccountSettings/__tests__/PersonalAccountSettings.test.tsx
@@ -13,7 +13,7 @@ import { i18n } from '/app/i18n'
 import { ACCESS_CONTROL_DISABLED_DOCUMENTATION_STATE } from '/app/local-resources/access-control/__fixtures__/documentationState'
 import { robotAuthReducer } from '/app/redux/robot-auth'
 
-import { PersonalAccountSettings } from '../PersonalAccountSettings'
+import { PersonalAccountSettings } from '..'
 
 import type { RenderResult } from '@testing-library/react'
 import type { AuthUserResponse, UpdateSelfRequest } from '@opentrons/api-client'
