+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/UserManagement/EditUserModal.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import { createPortal } from 'react-dom'
 import { useForm } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
@@ -10,7 +9,7 @@ import {
   SecondaryButton,
   WizardHeader,
 } from '@opentrons/components'
-import { useCreateUserMutation } from '@opentrons/react-api-client'
+import { useUpdateUserMutation } from '@opentrons/react-api-client'
 
 import { getTopPortalEl } from '/app/App/portal'
 import { useDocumentationState } from '/app/local-resources/access-control/useDocumentationState'
@@ -19,17 +18,13 @@ import { mapAuthUserMutationError } from '/app/resources/auth/mapAuthUserMutatio
 import {
   MANAGEABLE_USER_ACCOUNT_TYPES,
   USERNAME_MAX_LENGTH,
-} from './userAccount/constants'
-import { OneTimePasswordModal } from './userAccount/OneTimePasswordModal'
-import styles from './userAccount/userAccountForm.module.css'
-import { UserAccountIdentityFormFields } from './userAccount/UserAccountIdentityFormFields'
+} from '../userAccount/constants'
+import styles from '../userAccount/userAccountForm.module.css'
+import { UserAccountIdentityFormFields } from '../userAccount/UserAccountIdentityFormFields'
 
 import type { TFunction } from 'i18next'
 import type { JSX } from 'react'
-import type {
-  AuthUserAccountType,
-  CreateUserRequest,
-} from '@opentrons/api-client'
+import type { AuthUser, AuthUserAccountType } from '@opentrons/api-client'
 import type { DropdownOption } from '@opentrons/components'
 
 interface FormValues {
@@ -38,32 +33,31 @@ interface FormValues {
   accountType: AuthUserAccountType
 }
 
-export interface AddUserModalProps {
+export interface EditUserModalProps {
   robotName: string
+  user: AuthUser
   onClose: () => void
-  onUserCreated?: () => void
+  onUserUpdated?: () => void
 }
 
-export function AddUserModal({
+export function EditUserModal({
   robotName,
+  user,
   onClose,
-  onUserCreated,
-}: AddUserModalProps): JSX.Element {
+  onUserUpdated,
+}: EditUserModalProps): JSX.Element {
   const { t } = useTranslation(['device_settings', 'shared']) as {
     t: TFunction
   }
-  const [generatedPassword, setGeneratedPassword] = useState<string | null>(
-    null
-  )
   const documentationState = useDocumentationState(undefined, robotName)
-  const { createUser, isLoading: isSaving } =
-    useCreateUserMutation(documentationState)
+  const { updateUser, isLoading: isSaving } =
+    useUpdateUserMutation(documentationState)
   const { control, handleSubmit, watch, setValue, setError, clearErrors } =
     useForm<FormValues>({
       defaultValues: {
-        username: '',
-        fullName: '',
-        accountType: 'admin',
+        username: user.username,
+        fullName: user.fullName,
+        accountType: user.accountType,
       },
       mode: 'onBlur',
       reValidateMode: 'onChange',
@@ -71,50 +65,51 @@ export function AddUserModal({
 
   const { username, fullName, accountType } = watch()
   const accountTypeOptions: DropdownOption[] =
-    MANAGEABLE_USER_ACCOUNT_TYPES.map(accountType => ({
-      name: t(`desktop_user_role_${accountType}`),
-      value: accountType,
+    MANAGEABLE_USER_ACCOUNT_TYPES.map(type => ({
+      name: t(`desktop_user_role_${type}`),
+      value: type,
     }))
   const selectedAccountTypeOption =
     accountTypeOptions.find(option => option.value === accountType) ??
     accountTypeOptions[0]!
 
+  const trimmedUsername = username.trim()
+  const trimmedFullName = fullName.trim()
+  const hasChanges =
+    trimmedUsername !== user.username ||
+    trimmedFullName !== user.fullName ||
+    accountType !== user.accountType
+
   const isSaveDisabled =
     isSaving ||
-    username.trim() === '' ||
-    fullName.trim() === '' ||
-    username.length > USERNAME_MAX_LENGTH
+    !hasChanges ||
+    trimmedUsername === '' ||
+    trimmedFullName === '' ||
+    trimmedUsername.length > USERNAME_MAX_LENGTH
 
   const handleClose = (): void => {
     clearErrors()
-    setGeneratedPassword(null)
     onClose()
   }
 
-  const handleConfirm = (): void => {
-    onUserCreated?.()
-    handleClose()
-  }
-
   const onSubmit = (): void => {
-    const trimmedUsername = username.trim()
-    const request: CreateUserRequest = {
-      data: {
-        username: trimmedUsername,
-        fullName: fullName.trim(),
-        accountType,
+    void updateUser({
+      username: user.username,
+      request: {
+        data: {
+          ...(trimmedUsername !== user.username
+            ? { username: trimmedUsername }
+            : {}),
+          ...(trimmedFullName !== user.fullName
+            ? { fullName: trimmedFullName }
+            : {}),
+          ...(accountType !== user.accountType ? { accountType } : {}),
+        },
       },
-    }
-
-    void createUser(request)
-      .then(response => {
-        const { temporaryPassword } = response.data
-        if (temporaryPassword != null) {
-          setGeneratedPassword(temporaryPassword)
-          clearErrors()
-        } else {
-          handleClose()
-        }
+    })
+      .then(() => {
+        onUserUpdated?.()
+        handleClose()
       })
       .catch(error => {
         const formError = mapAuthUserMutationError<FormValues>(error, t)
@@ -124,23 +119,12 @@ export function AddUserModal({
       })
   }
 
-  if (generatedPassword != null) {
-    return (
-      <OneTimePasswordModal
-        password={generatedPassword}
-        message={t('desktop_add_user_success_message') as string}
-        onConfirm={handleConfirm}
-        onClose={handleClose}
-      />
-    )
-  }
-
   return createPortal(
     <ModalShell
       width="31.25rem"
       header={
         <WizardHeader
-          title={t('desktop_add_user')}
+          title={t('desktop_edit_user')}
           onExit={handleClose}
           hideStepText
           exitButtonCopy={t('shared:exit')}
@@ -176,7 +160,7 @@ export function AddUserModal({
                 {t('shared:cancel') as string}
               </SecondaryButton>
               <PrimaryButton type="submit" disabled={isSaveDisabled}>
-                {t('desktop_create_account') as string}
+                {t('shared:save') as string}
               </PrimaryButton>
             </div>
           </div>
